refactor: SankeyダイアグラムをDDDレイヤードアーキテクチャに移行
開発者がSankeyダイアグラム関連のコードを理解・変更しやすくするため、
utilsに分散していたロジックをDDDレイヤードアーキテクチャに整理。

- ドメインモデル追加: sankey-data.ts, category-aggregation.ts
- ドメインサービス追加: sankey-data-builder.ts
- Usecaseを public-finance コンテキストに移行
- types/sankey.ts をre-export形式に変更
- 不要になったutils/sankey-*.ts を削除
